added cors configuration and handled exceptions

+++ src/main/java/com/obgames/obgamesapi/service/CategoriaServiceImpl.java
@@ -3,14 +3,16 @@ package com.obgames.obgamesapi.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.obgames.obgamesapi.exceptions.ResourceNotFoundException;
+import com.obgames.obgamesapi.exceptions.ResponseStatusException;
+import com.obgames.obgamesapi.model.BrowserGame;
 import com.obgames.obgamesapi.model.Categoria;
+import com.obgames.obgamesapi.repository.BrowserGameRepo;
 import com.obgames.obgamesapi.repository.CategoriaRepo;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @Transactional
@@ -19,13 +21,16 @@ public class CategoriaServiceImpl implements CategoriaService {
     @Autowired
     private CategoriaRepo categoriaRepo;
 
+    @Autowired
+    private BrowserGameRepo browserGameRepo;
+
     @Override
     public Categoria createCategoria(Categoria categoria) {
         return categoriaRepo.save(categoria);
     }
 
     @Override
-    public Optional<Categoria> updateCategoria(Categoria categoria, String categoriaId) {
+    public Optional<Categoria> updateCategoria(Categoria categoria, String categoriaId) throws ResourceNotFoundException {
         Optional<Categoria> categoriaDb = this.categoriaRepo.findById(categoriaId);
         if (categoriaDb.isPresent()) {
             if (categoria.getId() == categoriaId) {
@@ -33,7 +38,7 @@ public class CategoriaServiceImpl implements CategoriaService {
                 return categoriaDb;
             }
         }
-        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Record not found with id : " + categoria.getId());
+        throw new ResourceNotFoundException("404", "Record not found with id : " + categoriaId);
 
     }
 
@@ -43,25 +48,32 @@ public class CategoriaServiceImpl implements CategoriaService {
     }
 
     @Override
-    public Categoria getCategoriaById(String categoriaId) {
+    public Categoria getCategoriaById(String categoriaId) throws ResourceNotFoundException {
 
         Optional<Categoria> categoriaDb = this.categoriaRepo.findById(categoriaId);
 
         if (categoriaDb.isPresent()) {
             return categoriaDb.get();
         } else {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Record not found with id : " + categoriaId);
+            throw new ResourceNotFoundException("404", "Record not found with id : " + categoriaId);
         }
     }
 
     @Override
-    public void deleteCategoria(String categoriaId) {
-        Optional<Categoria> categoriaDb = this.categoriaRepo.findById(categoriaId);
+    public void deleteCategoria(String categoriaId) throws ResourceNotFoundException, ResponseStatusException {
+        Categoria categoria = new Categoria();
+        categoria.setId(categoriaId);
+        List<BrowserGame> browserGames = this.browserGameRepo.findByCategoria(categoria);
+        if (browserGames.size() == 0){
+            Optional<Categoria> categoriaDb = this.categoriaRepo.findById(categoriaId);
 
-        if (categoriaDb.isPresent()) {
-            this.categoriaRepo.delete(categoriaDb.get());
+            if (categoriaDb.isPresent()) {
+                this.categoriaRepo.delete(categoriaDb.get());
+            } else {
+                throw new ResourceNotFoundException("404", "Record not found with id : " + categoriaId);
+            }
         } else {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Record not found with id : " + categoriaId);
+            throw new ResponseStatusException("5001","Essa categoria possui Browser Games vinculados. Para deletar, desvincule e tente novamente","Categoria ID: "+ categoriaId);
         }
 
     }
